Split http exception definitions
