+++ packages/server/src/app/models/errors/bad-request.exception.js
@@ -1,0 +1,9 @@
+const HttpException = require('./http.exception')
+
+class BadRequestException extends HttpException {
+  constructor(message) {
+    super(400, 'Bad Request', message)
+  }
+}
+
+module.exports = BadRequestException
