+++ packages/server/src/app/models/errors/not-found.exception.js
@@ -1,0 +1,9 @@
+const HttpException = require('./http.exception')
+
+class NotFoundException extends HttpException {
+  constructor(message) {
+    super(404, 'Not Found', message)
+  }
+}
+
+module.exports = NotFoundException
